+++ sdk/cosmos/azure-cosmos-tests/src/test/java/com/azure/cosmos/implementation/SessionTokenHelperTests.java
@@ -1,0 +1,36 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.azure.cosmos.implementation;
+
+import com.azure.cosmos.CosmosException;
+import org.mockito.Mockito;
+import org.testng.Assert;
+import org.testng.annotations.Test;
+
+import java.util.HashMap;
+
+import static com.azure.cosmos.implementation.TestUtils.mockDiagnosticsClientContext;
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class SessionTokenHelperTests {
+    @Test(groups = "unit")
+    public void missingPartitionKeyRangeIdInContext() {
+        try {
+            ISessionContainer mockedSessionContainer = Mockito.mock(ISessionContainer.class);
+
+            RxDocumentServiceRequest req =
+                RxDocumentServiceRequest.create(mockDiagnosticsClientContext(), OperationType.Read, ResourceType.Document,
+                    "/docs/",
+                    new Database(),
+                    new HashMap<>());
+
+            SessionTokenHelper.setPartitionLocalSessionToken(req, null, mockedSessionContainer);
+            Assert.fail("It should fail with internalServerError due to missing pkRangeId");
+        } catch (CosmosException cosmosException) {
+            assertThat(cosmosException.getStatusCode()).isEqualTo(HttpConstants.StatusCodes.INTERNAL_SERVER_ERROR);
+            assertThat(cosmosException.getSubStatusCode()).isEqualTo(HttpConstants.SubStatusCodes.MISSING_PARTITION_KEY_RANGE_ID_IN_CONTEXT);
+            assertThat(cosmosException.getMessage()).contains(RMResources.PartitionKeyRangeIdAbsentInContext);
+        }
+    }
+}